Using AbstractRepository in TopRepository
